feat: add utility constructors for ReactiveComponentExecutionMode
commit-id:fe5285a8
