+++ crates/starknet_integration_tests/src/config_utils.rs
@@ -1,22 +1,9 @@
 use std::fs::File;
 use std::io::Write;
-use std::net::SocketAddr;
 use std::path::PathBuf;
 
 use papyrus_config::dumping::{combine_config_map_and_pointers, SerializeConfig};
 use serde_json::{json, Map, Value};
-use starknet_sequencer_infra::component_definitions::{
-    LocalServerConfig,
-    RemoteClientConfig,
-    RemoteServerConfig,
-};
-use starknet_sequencer_infra::test_utils::get_available_socket;
-use starknet_sequencer_node::config::component_config::ComponentConfig;
-use starknet_sequencer_node::config::component_execution_config::{
-    ActiveComponentExecutionConfig,
-    ReactiveComponentExecutionConfig,
-    ReactiveComponentExecutionMode,
-};
 use starknet_sequencer_node::config::node_config::{
     SequencerNodeConfig,
     CONFIG_NON_POINTERS_WHITELIST,
@@ -108,68 +95,6 @@ fn strip_config_prefix(input: &str) -> &str {
         .strip_prefix("config.")
         .or_else(|| input.strip_prefix("required_params."))
         .unwrap_or(input)
-}
-
-// TODO(Nadin): Refactor the following functions to be static methods of
-// ReactiveComponentExecutionConfig.
-pub fn get_disabled_component_config() -> ReactiveComponentExecutionConfig {
-    ReactiveComponentExecutionConfig {
-        execution_mode: ReactiveComponentExecutionMode::Disabled,
-        local_server_config: None,
-        remote_client_config: None,
-        remote_server_config: None,
-    }
-}
-
-pub fn get_remote_component_config(socket: SocketAddr) -> ReactiveComponentExecutionConfig {
-    ReactiveComponentExecutionConfig {
-        execution_mode: ReactiveComponentExecutionMode::Remote,
-        local_server_config: None,
-        remote_client_config: Some(RemoteClientConfig { socket, ..RemoteClientConfig::default() }),
-        remote_server_config: None,
-    }
-}
-
-pub fn get_local_with_remote_enabled_component_config(
-    socket: SocketAddr,
-) -> ReactiveComponentExecutionConfig {
-    ReactiveComponentExecutionConfig {
-        execution_mode: ReactiveComponentExecutionMode::LocalExecutionWithRemoteEnabled,
-        local_server_config: Some(LocalServerConfig::default()),
-        remote_client_config: None,
-        remote_server_config: Some(RemoteServerConfig { socket }),
-    }
-}
-
-pub async fn get_http_only_component_config(gateway_socket: SocketAddr) -> ComponentConfig {
-    ComponentConfig {
-        http_server: ActiveComponentExecutionConfig::default(),
-        gateway: get_remote_component_config(gateway_socket),
-        monitoring_endpoint: Default::default(),
-        batcher: get_disabled_component_config(),
-        consensus_manager: ActiveComponentExecutionConfig::disabled(),
-        mempool: get_disabled_component_config(),
-        mempool_p2p: get_disabled_component_config(),
-        state_sync: get_disabled_component_config(),
-        l1_provider: get_disabled_component_config(),
-    }
-}
-
-pub async fn get_non_http_component_config(gateway_socket: SocketAddr) -> ComponentConfig {
-    ComponentConfig {
-        http_server: ActiveComponentExecutionConfig::disabled(),
-        monitoring_endpoint: Default::default(),
-        gateway: get_local_with_remote_enabled_component_config(gateway_socket),
-        ..ComponentConfig::default()
-    }
-}
-
-pub async fn get_remote_flow_test_config() -> Vec<ComponentConfig> {
-    let gateway_socket = get_available_socket().await;
-    vec![
-        get_http_only_component_config(gateway_socket).await,
-        get_non_http_component_config(gateway_socket).await,
-    ]
 }
 
 /// Transforms a nested JSON dictionary object into a simplified JSON dictionary object by
